feat(software): Rename software files for consistency - Rename software file names for consistency with other peripherals   (like iob-uart).
